add squareframelayout for name initial image
