+++ qabelbox/src/main/java/de/qabel/qabelbox/views/SquareFrameLayout.java
@@ -1,0 +1,48 @@
+package de.qabel.qabelbox.views;
+
+import android.annotation.TargetApi;
+import android.content.Context;
+import android.os.Build;
+import android.util.AttributeSet;
+import android.widget.FrameLayout;
+import android.widget.ImageView;
+
+/**
+ * Created by danny on 12.01.2016.
+ */
+public class SquareFrameLayout extends FrameLayout {
+
+    public SquareFrameLayout(Context context) {
+
+        super(context);
+    }
+
+    public SquareFrameLayout(Context context, AttributeSet attrs) {
+
+        super(context, attrs);
+    }
+
+    public SquareFrameLayout(Context context, AttributeSet attrs, int defStyleAttr) {
+
+        super(context, attrs, defStyleAttr);
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public SquareFrameLayout(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+
+        super(context, attrs, defStyleAttr, defStyleRes);
+    }
+
+    @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+
+        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
+        int width = getMeasuredWidth();
+        int height = getMeasuredHeight();
+
+        int dimension = Math.max(width, height);
+        setMeasuredDimension(dimension, dimension);
+
+
+    }
+}
