+++ qabelbox/src/main/res/layout/activity_create_identity.xml
@@ -30,7 +30,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
-            android:minHeight="200dp"
+            android:minHeight="220dp"
             android:orientation="vertical"/>
 
         <LinearLayout
